DJABBAMA-TECH
DJABBAMA TECH is a PWA( Progressive Web App which in French is “application web progressive) which uses artificial intelligence to facilitate the awareness of the Sahel region(more precisely in the Far North of Cameroon) through different well defined modules. The main aim is to sensitize the public on real time information and actualities, to promote the tourism and cultural sector, to be able to do online reservations(purchase of travelling tickets, hotel room preservations, and local products).